requirements.txt
streamlit
langchain
langchain-community
langchain-ollama
langchain-tavily
langgraph
pypdf
faiss-cpu
python-dotenv